Sorted dietary constraints and cuisines alphabetically on search page and added some padding to search page
